% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_map_colors}
\alias{generate_map_colors}
\title{Generate map colors}
\usage{
generate_map_colors(sf)
}
\arguments{
\item{sf}{An simple feature dataframe - the shapefile of investigation}
}
\value{
An integer vector - the indices of map colors
}
\description{
Generate map colors
}
