% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mex}
\alias{get_mex}
\title{Get the mex number of a vector}
\usage{
get_mex(v, colors, idx)
}
\arguments{
\item{v}{An logical vector - the intersection vector}

\item{colors}{An integer vector - the color assignment vector}

\item{idx}{An integer - the index}
}
\value{
An integer
}
\description{
Get the mex number of a vector
}
\examples{
get_mex(c(TRUE,TRUE,FALSE,FALSE,TRUE), 1:5, 4)

}
