% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_china}
\alias{plot_china}
\title{Main Plotting Function}
\usage{
plot_china(area = 11, palette = "Set1")
}
\arguments{
\item{area}{An integer vector - the area code of administraive divisions}

\item{palette}{A string - the palette used to color map}
}
\value{
A plot
}
\description{
Main Plotting Function
}
