% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china}
\alias{china}
\title{China administraive division shapefile}
\format{A simple feature dataframe of China administrative divisions}
\usage{
china
}
\description{
A simple feature dataframe of China administrative divisions
}
\keyword{datasets}
