% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_map_colors}
\alias{generate_map_colors}
\title{Generate map colors by greedy coloring algorithm so that bordering features are colored differently}
\usage{
generate_map_colors(sf)
}
\arguments{
\item{sf}{An simple feature dataframe - the shapefile of investigation}
}
\value{
An integer vector - the indices of map colors
}
\description{
Generate map colors by greedy coloring algorithm so that bordering features are colored differently
}
\examples{
generate_map_colors(head(china, 10))

}
