% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_china}
\alias{plot_china}
\title{Main Plotting Function}
\usage{
plot_china(
  area = 32,
  palette = "Set3",
  simplify_ratio = 0.05,
  prov_line = "solid",
  perf_line = "solid",
  county_line = "dashed"
)
}
\arguments{
\item{area}{An integer vector - the area code of administraive divisions. Default value is \code{32} for Jiangsu Province.}

\item{palette}{A string - the RColorBrewer palette used to color map. Default value is \code{"Set3"}}

\item{simplify_ratio}{A double - Between 0 and 1. The simplification ratio of map generation.}

\item{prov_line}{A string - The linetype of province-level divisions.}

\item{perf_line}{A string - The linetype of perfecture-level divisions.}

\item{county_line}{A string - The linetype of county-level divisions.}
}
\value{
A plot
}
\description{
Main Plotting Function
}
